requirements.txt
flask
openai
supabase
ultralytics
requests
pillow
gunicorn